+++ src/main/java/com/orzechowski/aidme/entities/rating/RatingRowMapper.java
@@ -11,6 +11,6 @@ public class RatingRowMapper implements RowMapper<Rating>
     public Rating mapRow(ResultSet rs, int rowNum) throws SQLException
     {
         return new Rating(rs.getLong(1), rs.getLong(2), rs.getString(3),
-                rs.getFloat(4));
+                rs.getInt(4));
     }
 }
